functioning version. Next steps are improving the strategymanager, and adjusting the probabilities

+++ internal/models/all_in.go
@@ -1,6 +1,6 @@
 package models
 
-func InvestDecisionMaking_allin(funds int) int {
+func InvestDecisionMaking_allin(funds float64) float64 {
 	// Implement your decision-making logic here
 	// This is a placeholder implementation
 	return funds
