+++ internal/models/default_half.go
@@ -1,6 +1,6 @@
 package models
 
-func InvestDecisionMaking_half(funds int, curround int, curscoreopo int) int {
+func InvestDecisionMaking_half(funds float64, curround int, curscoreopo int) float64 {
 
 	if curround == 15 || curround == 30 {
 		return funds // Invest half of the funds at round 15 and 30
